+++ service-worker.js
@@ -1,17 +1,56 @@
 ---
 layout: null
 ---
-'use strict';
-/* global init: readonly, config: readonly */
-/* {{ site.data.app.version | default: site.version }} */
-const CDN = 'https://cdn.kernvalley.us/';
+import { HermesWorker } from '{{ site.data.importmap.imports["@aegisjsproject/hermes/"] }}worker.js';
 
-try {
-	self.importScripts(new URL('./service-worker.js', CDN), '/sw-config.js');
+const staticDirs = ['js', 'css', 'img'];
 
-	if (init instanceof Function && typeof config === 'object') {
-		init(self, config);
-	}
-} catch(err) {
-	console.error(err);
-}
+new HermesWorker([
+	{
+		name: '{{ site.data.app.name | slugify }}',
+		version: '{{ site.data.app.version | default: site.version }}',
+		strategy: 'network-first',
+		pattern: new URLPattern({
+			baseURL: location.origin,
+			pathname: `/((?!(?:${staticDirs.join('|')})/).*)`
+		}),
+		prefetch: [
+			'/',
+			'/tags/kernville',
+			'/tags/lake-isabella',
+			'/tags/wofford-heights',
+			'/tags/mt-mesa',
+			'/tags/weldon',
+			'/tags/south-lake',
+			'/webapp.webmanifest',
+			'/create/',
+		].map(path => URL.parse(path, location.origin))
+	}, {
+		name: '{{ site.data.app.name | slugify }}-assets',
+		version: '{{ site.data.app.version | default: site.version }}',
+		strategy: 'stale-while-revalidate',
+		pattern: new URLPattern({
+			baseURL: location.origin,
+			pathname: `/(${staticDirs.join('|')})/*`,
+		}),
+		prefetch: [
+			'/js/index.min.js',
+			'/css/index.min.css',
+			'/img/icons.svg',
+			'/img/apple-touch-icon.png',
+			'/img/icon-192.png',
+			'/img/favicon.svg',
+		].map(path => URL.parse(path, location.origin))
+	}, {
+		name: 'unpkg',
+		strategy: 'cache-first',
+		pattern: new URLPattern({ baseURL: 'https://unpkg.com/', pathname: '/*' }),
+	}, {
+		name: 'imgur',
+		strategy: 'cache-first',
+		pattern: new URLPattern({
+			baseURL: 'https://i.imgur.com',
+			pathname: '/*',
+		}),
+	},
+]);