+++ CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Switch to `@aegisjsproject/hermes` for service worker
+
 ## [v2.1.6] - 2025-11-19
 
 ### Added